+++ lyrics-core/src/main/java/com/flipkart/lyrics/model/MetaInfo.java
@@ -12,17 +12,20 @@ public class MetaInfo {
     private final String className;
     private final String fullPackage;
     private final Map<String, TypeVariableName> genericVariables;
+    private boolean overrideAccessorGeneration;
 
     public MetaInfo(String className, String fullPackage, Map<String, TypeVariableName> genericVariables) {
         this.className = className;
         this.fullPackage = fullPackage;
         this.genericVariables = genericVariables;
+        this.overrideAccessorGeneration = false;
     }
 
     public MetaInfo(String className, String fullPackage) {
         this.className = className;
         this.fullPackage = fullPackage;
         this.genericVariables = new HashMap<>();
+        this.overrideAccessorGeneration = false;
     }
 
     public String getClassName() {
@@ -35,5 +38,13 @@ public class MetaInfo {
 
     public Map<String, TypeVariableName> getGenericVariables() {
         return genericVariables;
+    }
+
+    public boolean isOverrideAccessorGeneration() {
+        return overrideAccessorGeneration;
+    }
+
+    public void setOverrideAccessorGeneration(boolean overrideAccessorGeneration) {
+        this.overrideAccessorGeneration = overrideAccessorGeneration;
     }
 }